# Conflicts:
#	Sandbox/src/Level1.cpp
#	Zenith/include/Zenith/Graphics/Sprite.hpp
#	Zenith/include/Zenith/Math/Geometry.hpp
#	Zenith/include/Zenith/Math/Geometry.inl
#	Zenith/src/Graphics/Sprite.cpp

+++ Sandbox/src/Dragon.cpp
@@ -67,7 +67,7 @@ zth::u32 Dragon::next_frame()
     if (!_walking)
         timer.reset();
 
-    if (timer.elapsed_s() > _walking_animation_length)
+    if (timer.elapsed_s() > static_cast<double>(_walking_animation_length))
     {
         timer.reset();
         remainder = (remainder + 1) % 3;
